Add several ship types
Adding of three different types of IShip (Queen, Ufo, SmallShip). For the Ufo and SmallShip class the tick method hasn't been implemented yet

Adding of the shoot() method to the class Player and Handling of the SPACE key pressed event.

Add correction for the several type of ship correction and add of the bullet class for the shoots